new update  09-19-2017

+++ src/java/userBean/AddDocumentBean.java
@@ -4,6 +4,7 @@ import dao.InsertQueryDao;
 import dao.SelectQueryDao;
 import dbConnection.DBConnection;
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -15,6 +16,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.Random;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.servlet.ServletException;
@@ -57,6 +59,10 @@ public class AddDocumentBean extends HttpServlet {
     private ResultSet selectMaxId;
     private int documentId;
     private boolean addLetter;
+    private int count;
+    private File file1;
+    private String newFileName;
+    private File file;
 
     private String getFileName(final Part part) {
         final String partHeader = part.getHeader("content-disposition");
@@ -70,6 +76,27 @@ public class AddDocumentBean extends HttpServlet {
         return null;
     }
 
+    private void getRandom() throws FileNotFoundException, IOException {
+
+        Random random = new Random();
+        count = random.nextInt(999999999);
+        file1 = new File(uploaded_file + count + fileName);
+        newFileName = count + "_" + fileName;
+        if (file1.exists()) {
+            getRandom();
+        } else {
+            newFile.mkdir();
+            out1 = null;
+            out1 = new FileOutputStream(new File(uploaded_file + File.separator + newFileName));
+            fileContent = scanfile.getInputStream();
+            int read = 0;
+            final byte[] bytes = new byte[1024];
+            while ((read = fileContent.read(bytes)) != -1) {
+                out1.write(bytes, 0, read);
+            }
+        }
+    }
+
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
@@ -80,7 +107,7 @@ public class AddDocumentBean extends HttpServlet {
             date = new Date();
 
             photo = "";
-            uploaded_file = "E:/project/DMLC/DMLC/web/Uplopded_file";
+            uploaded_file = "E:/Programming/1. Office project/Project/DMLC/DMLC/web/Uplopded_file/";
             newFile = new File(uploaded_file);
 
             subject = new String(request.getParameter("subject").getBytes("ISO-8859-1"), "UTF-8");
@@ -94,6 +121,7 @@ public class AddDocumentBean extends HttpServlet {
             inputDate = dateFormat.format(date);
 
             fileName = getFileName(scanfile);
+            file = new File(uploaded_file + fileName);
 
             tableName = " document ";
             columnName = " date_time ";
@@ -107,20 +135,28 @@ public class AddDocumentBean extends HttpServlet {
             while (selectMaxId.next()) {
                 documentId = selectMaxId.getInt("document_id");
             }
-            tableName = " letter ";
-            columnName = " current_status, receiving_date, department_of_origin, request_id, subject_of_letter, end_date, document_id, short_desc, scan_file ";
-            values = "'" + status + "', '" + inputDate + "', '" + depOfOrigin + "', '" + requestId + "', '" + subject + "', '" + endDate + "', '" + documentId + "', '" + description + "', '" + fileName + "'";
-            addLetter = InsertQueryDao.insertQueryWithOutWhereClause(tableName, columnName, values);
 
-            newFile.mkdir();
-            out1 = null;
-            PrintWriter writer = response.getWriter();
-            out1 = new FileOutputStream(new File(uploaded_file + File.separator + fileName));
-            fileContent = scanfile.getInputStream();
-            int read = 0;
-            final byte[] bytes = new byte[1024];
-            while ((read = fileContent.read(bytes)) != -1) {
-                out1.write(bytes, 0, read);
+            if (file.exists()) {
+                getRandom();
+                tableName = " letter ";
+                columnName = " current_status, receiving_date, department_of_origin, request_id, subject_of_letter, end_date, document_id, short_desc, scan_file ";
+                values = "'" + status + "', '" + inputDate + "', '" + depOfOrigin + "', '" + requestId + "', '" + subject + "', '" + endDate + "', '" + documentId + "', '" + description + "', '" + newFileName + "'";
+                addLetter = InsertQueryDao.insertQueryWithOutWhereClause(tableName, columnName, values);
+            } else {
+                newFile.mkdir();
+                out1 = null;
+                PrintWriter writer = response.getWriter();
+                out1 = new FileOutputStream(new File(uploaded_file + File.separator + fileName));
+                fileContent = scanfile.getInputStream();
+                int read = 0;
+                final byte[] bytes = new byte[1024];
+                while ((read = fileContent.read(bytes)) != -1) {
+                    out1.write(bytes, 0, read);
+                }
+                tableName = " letter ";
+                columnName = " current_status, receiving_date, department_of_origin, request_id, subject_of_letter, end_date, document_id, short_desc, scan_file ";
+                values = "'" + status + "', '" + inputDate + "', '" + depOfOrigin + "', '" + requestId + "', '" + subject + "', '" + endDate + "', '" + documentId + "', '" + description + "', '" + fileName + "'";
+                addLetter = InsertQueryDao.insertQueryWithOutWhereClause(tableName, columnName, values);
             }
 
             if (addDocument) {
